[FEATURE] Add automatic gem release
This new GitHub Actions workflow will automatically publish the
gem to RubyGems when a new release is tagged with the `v` prefix,
e.g., `v5.0.0`.

For authentication, there needs to be a secret with the key
`RUBYGEMS_API_KEY` configured in the GitHub repository configuration.
The stored rubygems.org API key needs to have gem publication permissions,
and two-factor authentication for the corresponding rubygems.org account
should only be enabled for the UI, but must not be enabled for the API
(as there is no way to enter the second factor when the GitHub action
is run).